Convert the English documentation to ReStructuredText.
The image resources for the html files have been moved to not
be on a higher tree level where images for all subdirs of the
English docs are stored. Now the ones for a particular subdir
reside in the pictures/ directory inside of that subdir.

The content has not been updated, with few - if any - exceptions.
So if a link doesn't work, it probably doesn't work in the
original docs either. Fixing that kind of thing is a task for
general doc updating, not RST conversion.