avisynth_c: Fix compilation error after RC1 integration
First, AVS_ScriptEnvironment seems to have problems with how it's
being created in avisynth_c.cpp, which goes away if it's moved
back to avisynth_c.h.  This issue also caused problems in
PluginManager.cpp.

Second, the InterlockedIncrement call that was added at the same
time as moving avs_is_writable into the *.cpp causes another
compilation failure.

In both of these cases, the offending code has been commented out,
so that it can be accessed and fixed if necessary.