capi.h: partially protect against MSVC/GCC compatibility breakage
If a program that talks to the C-interface uses the GCC compliant
header, it becomes unable to use MSVC builds of AviSynth+. But
to build with GCC, those changes are necessary.

Guard against this by wrapping the change in a BUILDING_AVSCORE
ifdef block so that client programs will use the old calling
conventions.

Unfortunately, this comes with the caveat that 32-bit GCC builds
of AviSynth+ will not work with the client program. 64-bit GCC
builds are unaffected.