script.cpp: sprintf_s under MinGW-w64 requires Secure API
This is a quick-and-dirty way of working around this issue. A
better solution would be a header/source combination that defines
our own secure functions if Windows doesn't provide them.

The problem with MinGW-w64, though, is that the presence of the
Secure API causes other compatibility issues to be pervasive.