PluginManager.cpp: [MSVC|GCC] should only load [MSVC|GCC] plugins
The better way to accomplish this would be to detect it at runtime
and simply not autoload those plugins if the compiler for AviSynth
itself doesn't match it, but this is the next best option.

This commit simply tells GCC builds of AviSynth to use the value
of the PluginDir+GCC registry entry to find plugins, and ignore
PluginDir and PluginDir+.  Vice-versa for MSVC builds.

C plugins are an exception to this, since those can be loaded with
either MSVC- or GCC-built AviSynth+.  I'd suggest either having
duplicate copies, or using symbolic links for those.