CMakeLists.txt: Fix enabling of MSVC_CPU_ARCH option.
The if block it was in was an artifact of when the default for
x86 was SSE and for x86-64 was SSE2. The result was that the new
MSVC_CPU_ARCH option would only be effective for 32-bit builds,
and 64-bit builds would throw a warning about /arch: being an
empty command (although it still used SSE2 because MSVC has used
that as the default since 2012).

Now, the if logic is gone, making MSVC_CPU_ARCH effective for both
32-bit and 64-bit, and removing the warning.