CMakeLists.txt: prevent invalid settings for MSVC_CPU_ARCH on x64
MSVC doesn't allow setting /arch: to IA32, SSE, or SSE2 on x86-64,
so make sure that if the user passes one of those, we notify them
that the option will be ignored, and then remove the /arch: setting
from the build.

x86 builds are not affected at all, and it remains possible to set
MSVC_CPU_ARCH to AVX or AVX2 when building for x86-64.