version.h.in: Bump version number to 3.4
In the time since AviSynth+ was forked from classic, it has become
clear that some programs that would otherwise support AviSynth+
seamlessly end up choking on the versioning being so much lower
than classic.

This technically isn't AviSynth+'s fault, but there also tends to
be some confusion that arises based on what version of AviSynth+
to even use.  The version number perpetually being stuck at 0.1
isn't helping that situation, even if most focus is now on the
sequential revision.  The problem is that sequential revisions
are not helpful for management with pkg-config.

To that end, AviSynth+ should adopt 3.x as its major versioning.
This is, in fact, in keeping with the way previous versions of
classic AviSynth were versioned, in that v2.0 was the result of
some forks of v1.0 coalescing under new developers, and v2.5 was
a rewrite of v2.0 with more features.  Meanwhile, AviSynth3 was
poised to break compatibility with 2.5/2.6 entirely and use a
completely different scripting language, but it never got past an
extremely early alpha version that was never officially released.

AviSynth+, due to the extensive refactoring, should count as a new
major iteration, especially with all the new features (several of
which were on the v3 to-do list).  To help ease confusion with
that long-dead fork, the minor version is bumped to 4, with certain
milestones in AviSynth+'s development retroactively considered to
be the interim minor versions:

3.0 - (retro) AviSynth+ branching point, early October 2013
3.1 - (retro) 0.1 r1576, as official release, December 2013
3.2 - (retro) '0.2' r1827, MT, March 2015
3.3 - (retro) r2173, High bit depth & experimental GCC support,
August 2016
3.4 - upcoming, GCC support finalized, Clang support