avisynth.cpp: copy in AvxSynth's memory checking, as a test
On a related note, it turns out that AvxSynth's support for
then-OSX and BSD was isolated to just a couple spots in the code,
this memory function and a piece in the plugin manager (which
we aren't using).

This means that macOS and BSD support is mostly present
already.  A couple of future commits will enable them
in a more proper form.