Introduce generic AVS_POSIX define
In most cases, the use of AVS_LINUX throughout the code was really
true of all the other Unix-like systems, which meant for compiling
on macOS and BSD, we had to make sure AVS_LINUX was true. That was
weird, and it also ended up causing a bit of an issue with uchar.h
(seemingly a Linux-specific header) on macOS.

Another commit will actually change the name of the header to
posix.h to reflect the change.