Bump AVISYNTH_INTERFACE_VERSION to version 7.
In honesty, this probably should have been bumped a couple of times
already during AviSynth+ development, especially after high bit
depth was added, but it never was.

Basically, if API functions in avisynth.h and/or avisynth_c.h
change in a way that breaks compatibility with previous versions,
it needs to be bumped, and when new API functions are added, it
needs to be bumped.

While I'm not sure if any breaking changes to existing functions
have happened in Plus as of yet (due to the focus on maintaining
compatibility with 2.6), we certainly have added new API functions.

This means that checking avs_get_version could be used to ensure
the presence of Plus-specific functions, as version 7 will
guarantee it (and since classic 2.6 hasn't had a release or commit
in almost four years, avs_get_version returning 7 also means you're
using Plus).  Unfortunately, the lag here means that anything that
checks for version 6 compatibility will need to use GetProcAddress
or dlsym to figure out whether it's talking to classic or Plus.