avisynth.conf.in: fixup the autoload_directory path
CMAKE_INSTALL_FULL_LIBDIR is *supposed* to provide the absolute
path to the directory, but apparently it doesn't do that when
CMAKE_INSTALL_PREFIX hasn't been explicitly invoked by the user.

So if the user had omitted the CMAKE_INSTALL_PREFIX, expecting it
to be installed to /usr/local as is customary on many *nix systems,
avisynth.conf would instead say that the autoload_directory was
/lib/avisynth, not /usr/local/lib/avisynth.

So just do it the old way and use CMAKE_INSTALL_PREFIX and the
non-FULL CMAKE_INSTALL_LIBDIR to construct the absolute path.