TargetArch.cmake: add support for detecting ARMv8/aarch64
I'm not entirely sure if all of those predefined macros should
report as 'aarch64', since ARMv8 *does* support 32-bit, but for
the moment I think it's fairly safe to say that if we're supporting
compilation for the v8 features, we should assume aarch64.

I don't know how widespread AArch32 on ARMv8 support is on the OS
level as opposed to those OSes splitting between armv7/armel and
aarch64, or if it's even a distinction worth worrying about.