config.h: add basic detection for PowerPC
32-bit PPC was used in various game consoles, but most famously in
pre-Intel Macs. This form was also (technically or practically)
only big endian, in contrast to little endian, as used on x86(-64)
or the newer ARM chips that we care about.

64-bit PPC can be used in either big or little endian mode, although
as I don't have access to any of said hardware I can't say whether
that distinction matters for what config.h needs to know about.