DevIL_x86: don't store this with upx packing
It's consistently an issue when doing the installer packaging, and
doesn't always gets remembered.  If we store it in its original
unpacked form, this will no longer happen.

Honestly, we should be trying to move away from having binary .dlls
in the git repo at all, but until that happens this is one less
thing to have to worry about.